# book-social-network/book-netwok-api/src/main/java/com/ali/book/feedback/FeedbackResponse.java
package com.ali.book.feedback;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {

	private Double note;
	private String comment;
	private boolean ownFeedback;
}